third_party/WebKit/LayoutTests/http/tests/notifications/serviceworkerregistration-get-filter.html

<!doctype html>
<html>
  <head>
    <title>Notifications: ServiceWorkerRegistration.getNotifications() with a filter.</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src="../serviceworker/resources/test-helpers.js"></script>
    <script src="resources/test-helpers.js"></script>
  </head>
  <body>
    <script>
      // Tests that the getNotifications() function when used in a document returns
      // an array of the notifications which were previously displayed using the
      // same Service Worker registration id, having the same filter.
      async_test(function(test) {
          var scope = 'resources/scope/' + location.pathname,
              script = 'instrumentation-service-worker.js';

          testRunner.setPermission('notifications', 'granted', location.origin, location.origin);

          var registration = null;
          getActiveServiceWorkerWithMessagePort(test, script, scope).then(function(workerInfo) {
              registration = workerInfo.registration;

              return registration.showNotification('Hello, world!', {
                body: 'First notification',
                tag: 'banana',
              });
          }).then(function() {
              return registration.showNotification('Hello again, world!', {
                body: 'Second notification',
                tag: 'strawberry',
              });
          }).then(function() {
            return registration.getNotifications({ tag: 'strawberry' });
          }).then(function(notifications) {
              assert_equals(notifications.length, 1);

              assert_equals(notifications[0].title, 'Hello again, world!');
              assert_equals(notifications[0].body, 'Second notification');

              test.done();
          }).catch(unreached_rejection(test));

      }, 'ServiceWorkerRegistration.getNotifications() returns the opened notifications with the same filter.');
    </script>
  </body>
</html>
